renamed index page and components to portfolio
