+++ README.md
@@ -6,20 +6,29 @@ View the website live [here](https://dtran.tech/)!
 
 ### Getting Started
 
-First, pull environment variables from Vercel using the following command:
+This project needs to be set up using `yarn`. For some reason, `npm` is incapable of resolving peer dependencies,
+so you literally cannot install the packages unless you use `yarn`.
+
+To install `yarn`:
 
 ```sh
-npm run pull-env
+npm install --global yarn
 ```
 
-It will install the Vercel package if needed and then save the env variables to a `.env.local` file in the root folder.
+Now, we're ready to get going! First, pull environment variables from Vercel using the following command:
+
+```sh
+yarn run pull-env
+```
+
+It will install the Vercel package if needed and then save the env variables to a `.env` file in the root folder.
 
 ### Running the Server
 
 To run the development server:
 
 ```bash
-npm run dev
+yarn dev
 ```
 
 Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
