+++ src/app/projects/page.tsx
@@ -1,0 +1,14 @@
+import { Metadata } from "next";
+
+import { openGraph } from "../layout";
+
+import ProjectsPage from "./projects-page";
+
+export const metadata: Metadata = {
+  title: "Projects",
+  openGraph,
+};
+
+export default async function Page() {
+  return <ProjectsPage />;
+}
