+++ src/components/Resume/Subsection.tsx
@@ -4,7 +4,7 @@ import { FiChevronRight } from "react-icons/fi";
 
 import { TopLevelBlock } from "@contentful/rich-text-types";
 
-import { expandVariants } from "@/components/Index/Event";
+import { expandVariants } from "@/components/Portfolio/Event";
 import { convertDateToAbbrevString } from "@/utils/Date";
 import { ResumeSubsection } from "@/utils/types";
